When <Like/> is clicked, add article to likedArticles and color the button
This is a temporary solution since we don't have backend support. isInLikedArticles
selector, filters the likedArticles array to find if the article exists there.
If it does return true, so the button is colored blue. This creates a performance issue,
since the selector is run every time an article is rendered. To make this better,
Rerender library should be used to memoize the results.