Added user authorization process to redux
It does not yet communicate with the API, but the reducer can authorize
a user. To achieve this, two actions are needed, one that checks
if the user is authorized and one that authorizes the user. Later to be converted to a saga so it does communicate with backend to authorize the user.